analyse_barcodes: move pipeline code into new submodule 'barcodes.pipeline'.
